Add files via upload
Write an optimized resource allocator program that can be used for cost planning. The program takes the below 2 inputs:

Hours: No of hours the machine is required to run

Capacity: No of units are required (Will always be multiple of 10)

Based on these inputs, the program should optimally allocate the resources such that the cost of the capacity required is minimum. Calculate this for every region and show them as the below example.